work on bugs
