+++ Collections/PhoneBook.java
@@ -1,13 +1,26 @@
 package Collections;
 
-import java.util.Objects;
+import java.util.*;
 
 public class PhoneBook{
-    String surname;
-    String number;
 
-    public PhoneBook(String surname, String number) {
-        this.surname = surname;
-        this.number = number;
+    private final Map<String, Set<String>> phoneBooks = new HashMap<>();
+
+    private void addPhoneBookMap(String name, String number){
+
+//        Set<String> numbers = phoneBooks.getOrDefault(name, new HashSet<>());
+        if(phoneBooks.containsKey(name)){
+            phoneBooks.get(name).add(number);
+        }else {
+            Set<String> numbers = new HashSet<>();
+            numbers.add(number);
+            phoneBooks.put(name,numbers);
+        }
+        System.out.println(phoneBooks);
+    }
+
+    public Set<String> get(String name){
+        Map<String, Set<String>> phoneBooks = new HashMap<>();
+        return phoneBooks.getOrDefault(name, Collections.emptySet());
     }
 }
